+++ examples/advanced/src/gen/mocks/pet/index.ts
@@ -73,7 +73,6 @@ export {
   createUpdatePetWithFormStatus405Faker,
 } from './createUpdatePetWithFormFaker.ts'
 export {
-  createUploadFileDataFaker,
   createUploadFilePathPetIdFaker,
   createUploadFileQueryAdditionalMetadataFaker,
   createUploadFileResponseFaker,