+++ examples/advanced/src/gen/models/ts/index.ts
@@ -107,7 +107,6 @@ export type {
   UpdatePetWithFormStatus405,
 } from './pet/UpdatePetWithForm.ts'
 export type {
-  UploadFileData,
   UploadFilePathPetId,
   UploadFileQueryAdditionalMetadata,
   UploadFileRequestConfig,