+++ examples/advanced/src/gen/zod/pet/index.ts
@@ -73,7 +73,6 @@ export type {
   UpdatePetWithFormStatus405SchemaType,
 } from './updatePetWithFormSchema.ts'
 export type {
-  UploadFileDataSchemaType,
   UploadFilePathPetIdSchemaType,
   UploadFileQueryAdditionalMetadataSchemaType,
   UploadFileResponseSchemaType,
@@ -148,10 +147,4 @@ export {
   updatePetWithFormResponseSchema,
   updatePetWithFormStatus405Schema,
 } from './updatePetWithFormSchema.ts'
-export {
-  uploadFileDataSchema,
-  uploadFilePathPetIdSchema,
-  uploadFileQueryAdditionalMetadataSchema,
-  uploadFileResponseSchema,
-  uploadFileStatus200Schema,
-} from './uploadFileSchema.ts'
+export { uploadFilePathPetIdSchema, uploadFileQueryAdditionalMetadataSchema, uploadFileResponseSchema, uploadFileStatus200Schema } from './uploadFileSchema.ts'