+++ examples/client/src/gen2/models/UploadFile.ts
@@ -42,8 +42,6 @@ export type UploadFileJsonData = {
  */
 export type UploadFileFormData = {
   /**
-   * @description
-   * Format: `binary`
    * @type string
    */
   file: Blob