+++ examples/faker/src/gen/index.ts
@@ -178,7 +178,6 @@ export type {
   UpdateUserXmlData,
 } from './models/UpdateUser.ts'
 export type {
-  UploadFileData,
   UploadFilePathPetId,
   UploadFileQueryAdditionalMetadata,
   UploadFileRequestConfig,