+++ examples/faker/src/gen/models/index.ts
@@ -178,7 +178,6 @@ export type {
   UpdateUserXmlData,
 } from './UpdateUser.ts'
 export type {
-  UploadFileData,
   UploadFilePathPetId,
   UploadFileQueryAdditionalMetadata,
   UploadFileRequestConfig,