+++ examples/react-query/src/gen/index.ts
@@ -192,7 +192,6 @@ export type {
   UpdateUserXmlData,
 } from './models/UpdateUser.ts'
 export type {
-  UploadFileData,
   UploadFilePathPetId,
   UploadFileQueryAdditionalMetadata,
   UploadFileRequestConfig,