+++ examples/vue-query/src/gen/models/index.ts
@@ -177,7 +177,6 @@ export type {
   UpdateUserXmlData,
 } from './UpdateUser.ts'
 export type {
-  UploadFileData,
   UploadFilePathPetId,
   UploadFileQueryAdditionalMetadata,
   UploadFileRequestConfig,