+++ examples/zod/src/zod/zod/uploadFileSchema.ts
@@ -21,7 +21,3 @@ export type UploadFileStatus200SchemaType = z.infer<typeof uploadFileStatus200Sc
 export const uploadFileResponseSchema = uploadFileStatus200Schema
 
 export type UploadFileResponseSchemaType = z.infer<typeof uploadFileResponseSchema>
-
-export const uploadFileDataSchema = z.instanceof(File).optional()
-
-export type UploadFileDataSchemaType = z.infer<typeof uploadFileDataSchema>