+++ tests/3.0.x/__snapshots__/main/transformerStripDescriptions/types/Item.ts
@@ -4,26 +4,22 @@
 */
 
 /**
- * @example test
  * @type object
 */
 export type Item = {
     /**
      * @description
      * Format: `int64`
-     * @example test
      * @type integer
     */
     id: bigint;
     /**
-     * @example test
      * @type string
     */
     name: string;
     /**
      * @description
      * Format: `int32`
-     * @example test
      * @type integer | undefined
     */
     count?: number;