+++ tests/3.0.x/__snapshots__/pluginMcp/petStore/zod/addressSchema.ts
@@ -6,8 +6,8 @@
 import * as z from 'zod'
 
 export const addressSchema = z.object({
-  street: z.string().optional(),
-  city: z.string().optional(),
-  state: z.string().optional(),
-  zip: z.string().optional(),
+  street: z.string().optional().meta({ examples: ['437 Lytton'] }),
+  city: z.string().optional().meta({ examples: ['Palo Alto'] }),
+  state: z.string().optional().meta({ examples: ['CA'] }),
+  zip: z.string().optional().meta({ examples: ['94301'] }),
 })