+++ tests/3.0.x/__snapshots__/pluginZod/inferred/zod/orderSchema.ts
@@ -6,11 +6,11 @@
 import * as z from 'zod'
 
 export const orderSchema = z.object({
-  id: z.bigint().optional(),
-  petId: z.bigint().optional(),
-  quantity: z.int().optional(),
+  id: z.bigint().optional().meta({ examples: [10] }),
+  petId: z.bigint().optional().meta({ examples: [198772] }),
+  quantity: z.int().optional().meta({ examples: [7] }),
   shipDate: z.iso.datetime().optional(),
-  status: z.enum(['placed', 'approved', 'delivered']).optional().describe('Order Status'),
+  status: z.enum(['placed', 'approved', 'delivered']).optional().describe('Order Status').meta({ examples: ['approved'] }),
   complete: z.boolean().optional(),
 })