+++ tests/performance/CHANGELOG.md
@@ -1,5 +1,15 @@
 # performance
 
+## 0.0.1-beta.22
+
+### Patch Changes
+
+- Updated dependencies [[`0eabb97`](https://github.com/kubb-labs/plugins/commit/0eabb97b08188772b10a348c2144d51d0e4d6077), [`058c7ff`](https://github.com/kubb-labs/plugins/commit/058c7ffe6d8959c718543782f95d4f7bbef1cbe7), [`d0dc716`](https://github.com/kubb-labs/plugins/commit/d0dc716d746cd7c64d9b75597ebe9312ba51051d), [`1d58e1a`](https://github.com/kubb-labs/plugins/commit/1d58e1a6bf6514930bb8bbb214c6435bcb608530), [`0aa9573`](https://github.com/kubb-labs/plugins/commit/0aa9573825f6eff87e3301377016085ff334bc39), [`451f3b7`](https://github.com/kubb-labs/plugins/commit/451f3b7a24eb95fb4881bee8de59839e81686386), [`0aa9573`](https://github.com/kubb-labs/plugins/commit/0aa9573825f6eff87e3301377016085ff334bc39)]:
+  - @kubb/plugin-client@5.0.0-beta.64
+  - @kubb/plugin-ts@5.0.0-beta.64
+  - @kubb/plugin-faker@5.0.0-beta.64
+  - @kubb/plugin-zod@5.0.0-beta.64
+
 ## 0.0.1-beta.21
 
 ### Patch Changes